requirements.txt
Flask==2.3.3
ffmpeg-python==0.2.0
openai-whisper==20250625
requests==2.32.1
SQLAlchemy==2.0.22
PyPDF2==3.0.1
more-itertools==10.8.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
email-validator==2.1.1
Werkzeug==2.3.8
python-dotenv==1.0.1